requirements.txt
pytelegrambotapi~=4.24.0
requests~=2.31.0
python-dotenv~=1.0.1